+++ src/map.h
@@ -3,11 +3,11 @@
 
 struct
 {
-	uint16_t width, height, depth;
+	int width, height, depth;
 	struct tile *map;
 } map;
 
-extern void map_init( uint16_t width, uint16_t height, uint16_t depth );
+extern void map_init( int width, int height, int depth );
 
 //TODO replace this macro with some inline functions
 #define MAP(x, y, z) (map.map[z + x * map.depth + y * map.depth * map.width])
